requirements.txt
Pillow==9.3.0
python-slugify
Flask
imageio
ffmpeg-python
imageio-ffmpeg

